feat!: add a raw structure for the project
Add main.py to the project which checks if the program was run as a
script or imported as a module. If it was run as a script, it will
call the init and the loop functions. If it was imported as a module,
it will print a message and exit.